requirements.txt
Babel==2.2.0
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
Pygments==2.1.3
Sphinx==1.3.6
alabaster==0.7.7
appnope==0.1.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
cycler==0.10.0
decorator==4.0.9
docutils==0.12
functools32==3.2.3.post2
futures==3.0.5
gnureadline==6.3.3
ipykernel==4.3.1
ipyparallel==5.0.1
ipython==4.1.2
ipython-genutils==0.1.0
ipywidgets==4.1.1
jsonschema==2.5.1
jupyter-client==4.2.2
jupyter-core==4.1.0
matplotlib==1.5.1
mistune==0.7.2
nbconvert==4.1.0
nbformat==4.0.1
nose==1.3.1
notebook==4.1.0
numpy==1.10.4
pandas==0.18.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5.1
pyparsing==2.1.1
python-dateutil==2.5.1
pytz==2016.2
pyzmq==15.2.0
qtconsole==4.2.0
requests==2.9.1
scipy==0.13.3
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
sphinx-rtd-theme==0.1.9
terminado==0.6
testpath==0.3
tornado==4.3
traitlets==4.2.1
wsgiref==0.1.2
